+++ src/utils/stdio_safe_logger.ts
@@ -1,0 +1,4 @@
+export const stdioSafeLogger = Object.assign(Object.create(console), {
+  log: console.error.bind(console),
+  info: console.error.bind(console),
+}) as Console;